requirement.txt
pillow
pynput